Move Querys to Query.
