+++ src/Command/Command.php
@@ -16,7 +16,7 @@ use Yiisoft\Db\Connection\Connection;
 use Yiisoft\Db\Exception\InvalidArgumentException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Querys\Query;
+use Yiisoft\Db\Query\Query;
 use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Pdo\PdoValue;
