+++ src/Connection/Connection.php
@@ -15,7 +15,7 @@ use Yiisoft\Db\Exception\InvalidCallException;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
 use Yiisoft\Db\Factory\DatabaseFactory;
-use Yiisoft\Db\Querys\QueryBuilder;
+use Yiisoft\Db\Query\QueryBuilder;
 use Yiisoft\Db\Schemas\Schema;
 use Yiisoft\Db\Schemas\TableSchema;
 use Yiisoft\Db\Transactions\Transaction;
