+++ src/Expression/ExpressionBuilderTrait.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Expression;
 
-use Yiisoft\Db\Querys\QueryBuilder;
+use Yiisoft\Db\Query\QueryBuilder;
 
 /**
  * Trait ExpressionBuilderTrait provides common constructor for classes that should implement
