+++ src/Query/Conditions/ConjunctionCondition.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Querys\Conditions;
+namespace Yiisoft\Db\Query\Conditions;
 
 /**
  * Class ConjunctionCondition.
