+++ src/Query/Conditions/ExistsConditionBuilder.php
@@ -2,41 +2,34 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Querys\Conditions;
+namespace Yiisoft\Db\Query\Conditions;
 
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionBuilderTrait;
 use Yiisoft\Db\Expression\ExpressionInterface;
 
 /**
- * Class NotConditionBuilder builds objects of {@see NotCondition}.
+ * Class ExistsConditionBuilder builds objects of {@see ExistsCondition}.
  */
-class NotConditionBuilder implements ExpressionBuilderInterface
+class ExistsConditionBuilder implements ExpressionBuilderInterface
 {
     use ExpressionBuilderTrait;
 
     /**
      * Method builds the raw SQL from the $expression that will not be additionally escaped or quoted.
      *
-     * @param ExpressionInterface|NotCondition $expression the expression to be built.
+     * @param ExpressionInterface|ExistsCondition $expression the expression to be built.
      * @param array $params the binding parameters.
      *
      * @return string the raw SQL that will not be additionally escaped or quoted.
      */
     public function build(ExpressionInterface $expression, array &$params = []): string
     {
-        $operand = $expression->getCondition();
-        if ($operand === '') {
-            return '';
-        }
+        $operator = $expression->getOperator();
+        $query = $expression->getQuery();
 
-        $expession = $this->queryBuilder->buildCondition($operand, $params);
+        $sql = $this->queryBuilder->buildExpression($query, $params);
 
-        return "{$this->getNegationOperator()} ($expession)";
-    }
-
-    protected function getNegationOperator(): string
-    {
-        return 'NOT';
+        return "$operator $sql";
     }
 }
