+++ src/Query/Conditions/InConditionBuilder.php
@@ -2,12 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Querys\Conditions;
+namespace Yiisoft\Db\Query\Conditions;
 
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionBuilderTrait;
 use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Querys\Query;
+use Yiisoft\Db\Query\Query;
 
 /**
  * Class InConditionBuilder builds objects of {@see InCondition}.
