+++ src/Query/Conditions/NotConditionBuilder.php
@@ -2,34 +2,41 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Querys\Conditions;
+namespace Yiisoft\Db\Query\Conditions;
 
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionBuilderTrait;
 use Yiisoft\Db\Expression\ExpressionInterface;
 
 /**
- * Class ExistsConditionBuilder builds objects of {@see ExistsCondition}.
+ * Class NotConditionBuilder builds objects of {@see NotCondition}.
  */
-class ExistsConditionBuilder implements ExpressionBuilderInterface
+class NotConditionBuilder implements ExpressionBuilderInterface
 {
     use ExpressionBuilderTrait;
 
     /**
      * Method builds the raw SQL from the $expression that will not be additionally escaped or quoted.
      *
-     * @param ExpressionInterface|ExistsCondition $expression the expression to be built.
+     * @param ExpressionInterface|NotCondition $expression the expression to be built.
      * @param array $params the binding parameters.
      *
      * @return string the raw SQL that will not be additionally escaped or quoted.
      */
     public function build(ExpressionInterface $expression, array &$params = []): string
     {
-        $operator = $expression->getOperator();
-        $query = $expression->getQuery();
+        $operand = $expression->getCondition();
+        if ($operand === '') {
+            return '';
+        }
 
-        $sql = $this->queryBuilder->buildExpression($query, $params);
+        $expession = $this->queryBuilder->buildCondition($operand, $params);
 
-        return "$operator $sql";
+        return "{$this->getNegationOperator()} ($expession)";
+    }
+
+    protected function getNegationOperator(): string
+    {
+        return 'NOT';
     }
 }
