+++ src/Query/QueryBuilder.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Db\Querys;
+namespace Yiisoft\Db\Query;
 
 use Generator;
 use Yiisoft\Db\Connection\Connection;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
-use Yiisoft\Db\Querys\Conditions\ConditionInterface;
+use Yiisoft\Db\Query\Conditions\ConditionInterface;
 use Yiisoft\Db\Constraint\Constraint;
 use Yiisoft\Db\Constraint\ConstraintFinderInterface;
 use Yiisoft\Db\Exception\InvalidArgumentException;
@@ -17,8 +17,8 @@ use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Expression\ExpressionBuilder;
 use Yiisoft\Db\Expression\ExpressionBuilderInterface;
 use Yiisoft\Db\Expression\ExpressionInterface;
-use Yiisoft\Db\Querys\Conditions\HashCondition;
-use Yiisoft\Db\Querys\Conditions\SimpleCondition;
+use Yiisoft\Db\Query\Conditions\HashCondition;
+use Yiisoft\Db\Query\Conditions\SimpleCondition;
 use Yiisoft\Db\Schemas\ColumnSchemaBuilder;
 use Yiisoft\Db\Schemas\Schema;
 use Yiisoft\Db\Pdo\PdoValue;
@@ -324,8 +324,8 @@ class QueryBuilder
      *
      * @param string $table the table that new rows will be inserted into.
      * @param array|Query $columns the column data (name => value) to be inserted into the table or instance
-     * of {@see \Yiisoft\Db\Querys\Query|Query} to perform INSERT INTO ... SELECT SQL statement. Passing of
-     * {@see \Yiisoft\Db\Querys\Query|Query}.
+     * of {@see \Yiisoft\Db\Query\Query|Query} to perform INSERT INTO ... SELECT SQL statement. Passing of
+     * {@see \Yiisoft\Db\Query\Query|Query}.
      * @param array $params the binding parameters that will be generated by this method. They should be bound to the
      * DB command later.
      *
@@ -350,7 +350,7 @@ class QueryBuilder
      *
      * @param string $table the table that new rows will be inserted into.
      * @param array|Query $columns the column data (name => value) to be inserted into the table or instance of
-     * {\Yiisoft\Db\Querys\Query|Query} to perform INSERT INTO ... SELECT SQL statement.
+     * {\Yiisoft\Db\Query\Query|Query} to perform INSERT INTO ... SELECT SQL statement.
      * @param array $params the binding parameters that will be generated by this method.
      * They should be bound to the DB command later.
      *
