+++ src/Schemas/Schema.php
@@ -9,7 +9,7 @@ use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\IntegrityException;
 use Yiisoft\Db\Exception\InvalidCallException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Querys\QueryBuilder;
+use Yiisoft\Db\Query\QueryBuilder;
 use Yiisoft\Cache\CacheInterface;
 use Yiisoft\Cache\Dependency\TagDependency;
