+++ tests/QueryBuilderTest.php
@@ -7,9 +7,9 @@ namespace Yiisoft\Db\Tests;
 use Yiisoft\Db\Connection\Connection;
 use Yiisoft\Db\Mysql\QueryBuilder as MysqlQueryBuilder;
 use Yiisoft\Db\Pgsql\QueryBuilder as PgsqlQueryBuilder;
-use Yiisoft\Db\Querys\Query;
-use Yiisoft\Db\Querys\QueryBuilder;
-use Yiisoft\Db\Querys\Conditions\BetweenColumnsCondition;
+use Yiisoft\Db\Query\Query;
+use Yiisoft\Db\Query\QueryBuilder;
+use Yiisoft\Db\Query\Conditions\BetweenColumnsCondition;
 use Yiisoft\Db\Sqlite\Query\QueryBuilder as SqliteQueryBuilder;
 use Yiisoft\Db\Schemas\Schema;
 use Yiisoft\Db\Schemas\SchemaBuilderTrait;
